+++ packages/dara-core/tests/js/template-root.spec.tsx
@@ -31,22 +31,28 @@ describe('TemplateRoot', () => {
         expect(getByText('Frame, Menu')).toBeInstanceOf(HTMLSpanElement);
     });
 
-    it('should clean up cache on startup', () => {
+    it('should clean up cache on startup', async () => {
         const invalidKey = getSessionKey('SOME_OTHER_SESSION_KEY', 'test-uid-1');
         const validKey = getSessionKey('TEST_TOKEN', 'test-uid-2');
 
-        localStorage.setItem(getSessionKey('SOME_OTHER_SESSION_KEY', 'test-uid-1'), 'val1');
-        localStorage.setItem(getSessionKey('TEST_TOKEN', 'test-uid-2'), 'val2');
+        localStorage.setItem(invalidKey, 'val1');
+        localStorage.setItem(validKey, 'val2');
+        sessionStorage.setItem(invalidKey, 'val3');
+        sessionStorage.setItem(validKey, 'val4');
 
         expect(localStorage.getItem(invalidKey)).toEqual('val1');
         expect(localStorage.getItem(validKey)).toEqual('val2');
+        expect(sessionStorage.getItem(invalidKey)).toEqual('val3');
+        expect(sessionStorage.getItem(validKey)).toEqual('val4');
 
         wrappedRender(<TemplateRoot />);
 
         // Other session value should be cleaned up
-        waitFor(() => {
-            expect(localStorage.getItem(invalidKey)).toEqual(null);
+        await waitFor(() => {
+            expect(localStorage.getItem(invalidKey)).toEqual(undefined);
             expect(localStorage.getItem(validKey)).toEqual('val2');
+            expect(sessionStorage.getItem(invalidKey)).toEqual(undefined);
+            expect(sessionStorage.getItem(validKey)).toEqual('val4');
         });
     });
 });