+++ packages/dara-core/changelog.md
@@ -2,11 +2,14 @@
 title: Changelog
 ---
 
+## NEXT
+
+-   Fixed an issue where Local storage was not being cleared between sessions. 
+
 ## 1.5.0
 
 -   Internal (JS): added `RequestExtras` context to allow injecting additional e.g. headers into requests made by Dara in different parts of the component tree
 -   Move import discovery warnings to the `--debug` logger
--   Fixed an issue where Local storage was not being cleared between sessions. 
 
 ## 1.4.6