+++ packages/dara-core/changelog.md
@@ -2,7 +2,11 @@
 title: Changelog
 ---
 
-## NEXT
+## NEXT
+
+-   Fixed an issue where Local storage was not being cleared between sessions. 
+
+## 1.5.0
 
 -   Internal (JS): added `RequestExtras` context to allow injecting additional e.g. headers into requests made by Dara in different parts of the component tree
 -   Move import discovery warnings to the `--debug` logger